done with lab12

+++ lab12.py
@@ -7,7 +7,7 @@ import time
 import re
 import time
 
-specLab = sys.argv[1]
+specLab = "lab12"
 
 class Checker():
 
@@ -17,7 +17,7 @@ class Checker():
         f = open("tests/" + self.labName + ".txt", "r")
         txt = json.loads(f.read())
         self.files = txt["files"]
-        self.required_file = txt["required"]
+        self.required_file = txt["required"][0]
         self.expected_output = txt["output"]
 
     def getAssignmentList(self):
@@ -29,52 +29,33 @@ class Checker():
     
 
     def compileCheck(self, originalFiles):
-        print "compileCheck"
-        print originalFiles
-        for f in originalFiles:
-            print "f is " + f
-            for fName in self.files:
-                print "fName is " + fName
-                cleanedUpFileName = f.split("_")[len(f.split("_")) -1]
-                cleanedUpFileName = cleanedUpFileName.split("-")[0]
+        if self.files[0].replace(".java", "") in originalFiles:
+            cleanedUpFileName = originalFiles.split("_")[-1]
 
-                if ".java" not in cleanedUpFileName:
-                    cleanedUpFileName += ".java"
-                print "cleanedUpFileName " + cleanedUpFileName 
-                if fName in cleanedUpFileName or fName.lower() in cleanedUpFileName:
-                    print "\tfName in f"
-                    subprocess.call("cp submissions/" + self.labName + "/" + f + " sandbox/", shell=True)
-                    subprocess.call("mv sandbox/" + f + " sandbox/" + fName, shell=True)
+            if cleanedUpFileName != self.files[0]:
+                cleanedUpFileName = self.files[0]
+
+            if ".java" not in cleanedUpFileName:
+                cleanedUpFileName += ".java"
+
+            subprocess.call("cp submissions/" + self.labName + "/" + originalFiles + " sandbox/", shell=True)
+            subprocess.call("mv sandbox/" + originalFiles + " sandbox/" + self.files[0], shell=True)
         
-        temp = subprocess.check_output("ls sandbox", shell=True)
-        print "ls sandbox\n" + str(temp)
-
-        subprocess.call("cp requiredClasses/" + self.required_file + " sandbox/", shell=True)
-        self.helperClasses()
-        subprocess.call("javac -encoding ISO-8859-1 sandbox/*.java", shell=True)
-        print "done compileCheck"
+            subprocess.call("cp requiredClasses/" + self.required_file + " sandbox/", shell=True)
+            subprocess.call("javac -encoding ISO-8859-1 sandbox/*.java", shell=True)
         
-    def compare(self, stringA, stringB):    
-        difference = [li for li in list(difflib.ndiff(stringA, stringB)) if li[0] != ' ']
-
-        return difference
-
-    def checkOutputGrade(self, output):
-        #index = output.index("Grade=" ) + len("Grade=")
-        index = output.index("Total:")
-        grade = output[index:]
-        return grade
-
     def outputCheck(self):
-        temp = subprocess.check_output("ls sandbox", shell=True)
-        print "ls sandbox\n" + str(temp)
-        results = subprocess.check_output("cd sandbox/ && java " + self.required_file.replace(".java", ""), shell=True)
-        results = results.replace("\n", "").replace("\t", "").strip().replace(" ", "")
-        #listOfErrors = [x for x in results.split("\n") if "Failed" in x]
-        #numOfErrors = len(listOfErrors)
-        subprocess.call("rm sandbox/*", shell=True)
+        results = ""
+        try:
+            results = subprocess.check_output("cd sandbox/ && java " + self.required_file.replace(".java", ""), shell=True)
+            results = results.split("\n")
+            results = results[:-1]
+            results = results[-1]
+            results = ["score", results]
+        except Exception as e:
+            results = ["error", str(e)]
+
         return results
-        #return {"numOfErrors": numOfErrors, "errorList": str(listOfErrors) }
 
     def generateResults(self, ob):
         with open("results/" + self.labName + ".csv", "wb") as csvfile:
@@ -87,26 +68,13 @@ class Checker():
 
     def run(self):
         assignmentList = self.getAssignmentList()
-        assignmentList_new = []
-        i = 0
-        count = 0
-        step = len(self.files)
-        numOfSections = len(assignmentList) / step
-        while count < numOfSections:
-            section = [assignmentList[i], assignmentList[i + 1], assignmentList[i + 2]]
-            assignmentList_new.append(section)
-            i += step
-            count += 1
 
         results = {}
         i = 0
-        for assignment in assignmentList_new:
-        #for assignment in assignmentList:
+        for assignment in assignmentList:
             internal_ob = {}
-            aFile = str(assignment[0])
-            internal_ob["name"] = aFile[:aFile.index('_')]
-            print "student " + internal_ob["name"]
-            #internal_ob["name"] = assignment[0].split("_")[0]
+            internal_ob["name"] = assignment.split("_")[0]
+            print internal_ob["name"]
             internal_ob["output"] = ""
             internal_ob["grade"] = 0
             internal_ob["errors"] = ""
@@ -115,14 +83,17 @@ class Checker():
             try:
                 self.compileCheck(assignment)
             except Exception as e:
-                print "\tfailed compileCheck"
-                internal_ob["errors"] = str(e)
-            try:
-                internal_ob["output"] = self.outputCheck()
-                internal_ob["grade"] = str(self.checkOutputGrade(internal_ob["output"]))
-            except Exception as e:
-                print "\tfailed outputCheck"
-                internal_ob["errorList"] = str(e)
+                internal_ob["errors"] = "failed compileCheck"
+
+            output = self.outputCheck()
+
+            if output[0] == "score":
+                internal_ob["output"] = output[1]
+                print internal_ob["output"]
+                internal_ob["grade"] = internal_ob["output"].split("/")[0].replace("Total:", "").strip()
+                print internal_ob["grade"]
+            else:
+                internal_ob["errorList"] = output[1]
                 
             subprocess.call("rm sandbox/*", shell=True)
             results[i] = internal_ob
